fix: Adds look vector correction code
This is needed for the Xreal line of devices as the reverse engineered
driver works, but has... issues. One of them being normal sensor drift,
the other being ~2330 units in any direction for some reason.

This is a hack but I can't think of a normal person that can spin their
head around fast enough to trigger anything past 7 units/pull cycle,
especially considering our already high pulling rate (basically as
fast as we can)